Move PostgreSQL-specific methods to PostgreSql92Dialect. Fix testLikeOperator() when scopes with different dialects are present.
